Prepend generated data to yayates data
Take the generated data that we created and just prepend the data that happens before 2010 to the Yayates data. I noticed some errors in the Yayates data though, so this will have to be updated to have our corrected data and use the Yayates data where we couldn't generate the data.